requirements.txt
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
ccxt==4.1.81
pandas==2.1.1
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1
pydantic==2.10.2
